pyproject.toml
#####################################################################################################
# general
#####################################################################################################

[project]
name = "fastapi-project"
version = "0.1.0"
description = "FastAPI Интернет-магазин"
readme = "README.md"
authors = [
    {name = "Dmitry Zhigalo"}
]
requires-python = ">=3.13"

#####################################################################################################
# dependencies
#####################################################################################################

dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.121.3",
    "greenlet>=3.2.4",
    "orjson>=3.11.4",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
]

[dependency-groups]
lint = [
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "poethepoet>=0.37.0",
    "ruff>=0.12.10",
]

#####################################################################################################
# ruff
#####################################################################################################

[tool.ruff]
line-length = 150
target-version = "py313"
cache-dir = "~/.cache/ruff"
preview = true
exclude = [
  ".git",
  ".venv",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88
quote-style = "single"

[tool.ruff.lint]
select = [
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "COM",    # flake8-commas
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "N",      # pep8-naming
    "N",      # pep8-naming
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "PL",     # pylint
    "F",      # pyflakes
    "UP",     # pyupgrade
    "RUF",    # ruff-specific rules
    "TRY",    # tryceratops
    "D101",   # undocumented-public-class
    "D103",   # undocumented-public-function
]
allowed-confusables = ["е", "с", "о", "г"]
ignore = [
    "E302", # blank-lines-top-level
    "E305", # blank-lines-after-function-or-class
]

[tool.ruff.lint.per-file-ignores]
"src/api/auth.py" = ["E712", "B008", "RUF029"]
"src/api/main.py" = ["PLW1508"]
"src/models/*.py" = ["D101", "F821"]
"src/routes/*.py" = ["E712", "B008"]
"src/utils/routes.py" = ["E712"]
"src/migrations/env.py" = ["D103", "E712"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

#####################################################################################################
# isort
#####################################################################################################

[tool.isort]
py_version = 313
line_length = 120
combine_as_imports = true
combine_star = true
lexicographical = true
src_paths = ['src']
remove_redundant_aliases = false
include_trailing_comma = true
color_output = true
multi_line_output = 3

#####################################################################################################
# mypy
#####################################################################################################

[tool.mypy]
cache_dir = '~/.cache/.mypy'
python_version = '3.13'
platform = 'linux'
namespace_packages = true
explicit_package_bases = true
plugins = [
    'pydantic.mypy',
]

follow_imports = "normal"
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

strict_equality = true
strict_concatenate = true
strict_optional = true

strict = true
implicit_reexport = true

show_error_codes = true
show_error_context = true
show_column_numbers = true
pretty = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

#####################################################################################################
# poe
#####################################################################################################

[tool.poe.tasks]
isort = "isort --jobs -1 ."
mypy = { shell = "mypy src || true" }
all = ['isort', 'mypy']
